Update atts generator to cast to 'int' the return value of GetNum* method, regenerated atts
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@21609 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/common/state/ColorTableAttributes.C
@@ -875,7 +875,7 @@ ColorTableAttributes::RemoveColorTables(int index)
 int
 ColorTableAttributes::GetNumColorTables() const
 {
-    return colorTables.size();
+    return (int)colorTables.size();
 }
 
 // ****************************************************************************
